// src/shared/interfaces/stats-type.interface.ts
export enum StatsType {
  KILLS = 'kills',
  DEATHS = 'deaths',
  KD = 'kd',
  MAX_STREAK = 'maxstreak',
  ELO = 'elo',
  LEVEL = 'level',
  KOTH = 'koth',
}

export enum StatsTypeSurvival {
  KILLS = 'kills',
  DEATHS = 'deaths',
  KD = 'kd',
  MAX_STREAK = 'maxstreak',
  ELO = 'elo',
  KOTH = 'koth',
}